Move ServerGroup and group_servers from ui.search to server
To avoid unnecessary dependency of CLI on Gtk.
